supermarket_extractions/requirements.txt
requests
beautifulsoup4
pyyaml
psycopg2